Added candidate and recruiter profile and job post flow

+++ tests/setup.js
@@ -1,13 +1,20 @@
 const mongoose = require("mongoose");
 
 beforeAll(async () => {
-    await mongoose.connect(process.env.MONGO_URI_TEST, {
-        useNewUrlParser: true,
-        useUnifiedTopology: true
-    });
+    await mongoose.connect(process.env.MONGO_URI_TEST);
+});
+
+afterEach(async () => {
+    const collections = Object.keys(mongoose.connection.collections);
+    for (const collectionName of collections) {
+        const collection = mongoose.connection.collections[collectionName];
+        await collection.deleteMany({});
+    }
 });
 
 afterAll(async () => {
-    await mongoose.connection.dropDatabase();
+    if (mongoose.connection.readyState === 1) {
+        await mongoose.connection.dropDatabase();
+    }
     await mongoose.connection.close();
 });
